+++ bonus/tests/test_is_sorted.c
@@ -6,13 +6,13 @@
 */
 
 #include <criterion/criterion.h>
-#include "../include/pushswap.h"
+#include "pushswap.h"
 
 Test(is_sorted, sorted_case)
 {
-    int list1[4] = {0, 4, 85, 5604};
-    int list2[2] = {456, 457};
-    int list3[3] = {52, 52, 52};
+    float list1[4] = {0, 4, 85, 5604};
+    float list2[2] = {456, 457};
+    float list3[3] = {52, 52, 52};
 
     cr_expect_eq(is_sorted(list1, 4), 1);
     cr_expect_eq(is_sorted(list2, 2), 1);
@@ -21,11 +21,11 @@ Test(is_sorted, sorted_case)
 
 Test(is_sorted, not_sorted_case)
 {
-    int list1[5] = {0, 4, 85, 5604, 5603};
-    int list2[2] = {458, 457};
-    int list3[3] = {52, 52, 51};
-    int list4[5] = {0, 4, 5604, 85, 5603};
-    int list5[5] = {4, 3, 85, 5604, 5603};
+    float list1[5] = {0, 4, 85, 5604, 5603};
+    float list2[2] = {458, 457};
+    float list3[3] = {52, 52, 51};
+    float list4[5] = {0, 4, 5604, 85, 5603};
+    float list5[5] = {4, 3, 85, 5604, 5603};
 
     cr_expect_eq(is_sorted(list1, 5), 0);
     cr_expect_eq(is_sorted(list2, 2), 0);
@@ -36,7 +36,7 @@ Test(is_sorted, not_sorted_case)
 
 Test(is_sorted, one_number_case)
 {
-    int list1[1] = {56};
+    float list1[1] = {56};
 
     cr_expect_eq(is_sorted(list1, 1), 1);
 }
