+++ bonus/tests/test_pushswap_ra.c
@@ -7,14 +7,14 @@
 
 #include <criterion/criterion.h>
 #include <criterion/redirect.h>
-#include "../include/pushswap.h"
+#include "pushswap.h"
 
 Test(pushswap_ra, several_nbrs_in_list, .init=cr_redirect_stdout)
 {
-    int actual1[3] = {0, 1, 4};
-    int actual2[2] = {0, 1};
-    int expected1[3] = {1, 4, 0};
-    int expected2[2] = {1, 0};
+    float actual1[3] = {0, 1, 4};
+    float actual2[2] = {0, 1};
+    float expected1[3] = {1, 4, 0};
+    float expected2[2] = {1, 0};
 
     pushswap_ra(actual1, 3);
     pushswap_ra(actual2, 2);
@@ -25,7 +25,7 @@ Test(pushswap_ra, several_nbrs_in_list, .init=cr_redirect_stdout)
 
 Test(pushswap_ra, one_nbr_in_list, .init=cr_redirect_stdout)
 {
-    int actual[1] = {0};
+    float actual[1] = {0};
 
     pushswap_ra(actual, 1);
     cr_assert_arr_eq(actual, actual, 1);
